+++ airflow/auth/managers/simple/routes/login.py
@@ -22,7 +22,7 @@ from fastapi import status
 from airflow.api_fastapi.common.router import AirflowRouter
 from airflow.api_fastapi.core_api.openapi.exceptions import create_openapi_http_exception_doc
 from airflow.auth.managers.simple.datamodels.login import LoginBody, LoginResponse
-from airflow.auth.managers.simple.services.simple_auth_manager_login import SimpleAuthManagerLogin
+from airflow.auth.managers.simple.services.login import SimpleAuthManagerLogin
 from airflow.configuration import conf
 
 login_router = AirflowRouter(tags=["SimpleAuthManagerLogin"])
